pyproject.toml
[project]
name = "mlsc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "ruff>=0.14.14",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

[project.scripts]
mlsc = "mlsc.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mlsc"]
